backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
azure-cognitiveservices-speech